A first version of PersistentObjectDStore. Equivalent with PersistentObjectStore, this class adds support for the persistenceId as store identifier.
Different from PersistentObjectStore, we do no extend StoreOfStateful. First, since the events in dstore are different from store, another implementation would be necessary. Secondly, StoreOfStateful was not a great success, and we should think about not doing this anymore.

This would constitute a new major version, but not a new mark version, as it is an extra class, and thus a backward compatible new feature. Notably is that other code, like CrudDao, doesn't work with dstores yet.