Making polymorphAmdRevive dependent on CrudDao was stupid and unnecessary. It only needs a _Cache. Changed the dependency from a CrudDao to a _Cache.
There is one method that is not named exactly the same: crudDao.getCachedByTypeAndId goes to cache.getByTypeAndId.
We do not change the method names, but also in all calls to polymorphAmdRevive now give the cache, and not the CrudDao.
This is simple: the revive method of CrudDao now has 1 argument less (always the self, intended to be the cache). Instead, realisations of this abstract method should give this._cache to polymorphAmdRevive.

This is an incompatible change! Mark version changed!