The cache of the CrudDao (a _Cache instance) is now public, and must be injected (IoC).
The reason is that we now have different code that must share a common _Cache instance.
The property _cache is renamed to cache, and it is no longer instantiated in the constructor. Therefor, we also no longer need a dependency on _Cache in CrudDao.

This is an incompatible change, and requires a new mark version.