Bugfix: missing this.inherited(arguments) call in PersistentObjectButtonEditPane; BUT this reveals other bugs!
By missing this inherited call, we did never apply the css style "SemanticObjectPane" on the pane. Now we do, and we see that this reveals a load of missing features we once added, but never got working. However, we also now loose margins, and have strange focus behaviour.

We combine this with an override in the immediate superclass _PersistentObjectEditPane, where we conveniently "forget" the call too for now.