Rework retrieve() to fix a giant bug when we get an IdNotFoundException. Still we get too much error reporting from "somewhere", and it doesn't work on update (POST).
Now we use _cleanupAfterRemove.
Also, we forget the retrieve promise after revive, not before.

_PersistentObjectEditPane now listens to the topic (like TooMany lists), and reacts by closing itself on IdNotFoundException.
Note that an alert is impossible, because it blocks everything. We need a modal communication mechanism that doesn't block JS, or to use flash messages (see IDEA).
refresh() now reports the error as info
cancel() and remove() and save() (via _handleSafeException()) now do not close themselves on IdNotFoundException, but leaves that to the topic event handling

We might want to do the same thing for SecurityException one day.

Also change i18n is message, but the message is not used anymore, now (until we get a non-blocking mechanism).