We now signal each (instance) ActionCompleted on a topic, with enough information.
There is a lot of text in this commit, which is not really good documentation, but the reasoning to get to this result.
Reworked _poAction, retrieve, remove
to handle IdNotFoundException correctly, and signal action completion on the topic. For this, also cleanupAfterRemove was changed. Amongst other things, in this case, events are now later, after the objects have been changed to a null peristenceId, and after related objects are refreshed too. This order is not nice.

_PersistentObjectEditPane now handles the new type of events.

This is a big refactor. It behaves nicely on first tests, but there are still some problems.
Now, ToManyList might trigger an update for something that is already deleted, and we know it.
Also, I see too much retrieves (cancels?).

There are still some MUDO's and TODO's which need to be cleaned up.